Move summarizer types to new file
