+++ packages/runtime/container-runtime/src/summaryManager.ts
@@ -11,7 +11,8 @@ import { IFluidObject, IRequest } from "@fluidframework/core-interfaces";
 import { IContainerContext, LoaderHeader } from "@fluidframework/container-definitions";
 import { ISequencedClient } from "@fluidframework/protocol-definitions";
 import { DriverHeader } from "@fluidframework/driver-definitions";
-import { ISummarizer, createSummarizingWarning, ISummarizingWarning, SummarizerStopReason } from "./summarizer";
+import { createSummarizingWarning } from "./summarizer";
+import { ISummarizer, ISummarizingWarning, SummarizerStopReason } from "./summarizerTypes";
 import { SummarizerClientElection, summarizerClientType } from "./summarizerClientElection";
 import { Throttler } from "./throttler";
