+++ packages/runtime/container-runtime/src/test/runningSummarizer.spec.ts
@@ -15,7 +15,7 @@ import {
 } from "@fluidframework/protocol-definitions";
 import { MockDeltaManager, MockLogger } from "@fluidframework/test-runtime-utils";
 import { RunningSummarizer } from "../runningSummarizer";
-import { SummarizerStopReason } from "../summarizer";
+import { SummarizerStopReason } from "../summarizerTypes";
 import { SummaryCollection } from "../summaryCollection";
 
 describe("Runtime", () => {
